backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
pydantic[email]>=2.7.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
qrcode[pil]>=7.4.2
Pillow>=10.0.0
python-dotenv>=1.0.1
aiofiles>=23.2.1
bcrypt>=4.0.0
openpyxl>=3.1.2
reportlab>=4.2.0
httpx>=0.27.0
